+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderStockServiceImpl.java
@@ -254,17 +254,16 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService {
     if (this.isStockMoveProduct(saleOrderLine)) {
 
       Unit unit = saleOrderLine.getProduct().getUnit();
-      BigDecimal priceDiscounted = saleOrderLine.getPriceDiscounted();
-      BigDecimal companyUnitPriceUntaxed = saleOrderLine.getProduct().getCostPrice();
+      BigDecimal valuatedUnitPrice = saleOrderLine.getProduct().getCostPrice();
       if (unit != null && !unit.equals(saleOrderLine.getUnit())) {
         qty =
             unitConversionService.convert(
                 saleOrderLine.getUnit(), unit, qty, qty.scale(), saleOrderLine.getProduct());
-        priceDiscounted =
+        valuatedUnitPrice =
             unitConversionService.convert(
                 unit,
                 saleOrderLine.getUnit(),
-                priceDiscounted,
+                valuatedUnitPrice,
                 appBaseService.getNbDecimalDigitForUnitPrice(),
                 saleOrderLine.getProduct());
       }
@@ -275,7 +274,7 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService {
         taxRate = taxLine.getValue();
       }
       if (saleOrderLine.getQty() != BigDecimal.ZERO) {
-        companyUnitPriceUntaxed =
+        valuatedUnitPrice =
             saleOrderLine
                 .getCompanyExTaxTotal()
                 .divide(
@@ -290,8 +289,7 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService {
               saleOrderLine.getProductName(),
               saleOrderLine.getDescription(),
               qty,
-              priceDiscounted,
-              companyUnitPriceUntaxed,
+              valuatedUnitPrice,
               unit,
               stockMove,
               StockMoveLineService.TYPE_SALES,
@@ -314,7 +312,6 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService {
               null,
               saleOrderLine.getProductName(),
               saleOrderLine.getDescription(),
-              BigDecimal.ZERO,
               BigDecimal.ZERO,
               BigDecimal.ZERO,
               null,